backend/requirements.txt
flask==3.1.*
flask-sqlalchemy==3.1.*
flask-cors==5.0.*
flask-jwt-extended==4.7.*
python-dotenv==1.0.*
gunicorn==23.0.*
